% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/FizzBuzz.R
\name{FizzBuzz}
\alias{FizzBuzz}
\title{FizzBuzz}
\usage{
FizzBuzz(n)
}
\description{
This function accepts a vector of integers. It replaces all integers
 divisible by 3 with the word "Fizz", all integers divisible by 5
 with the word "Buzz", and all integers divisible by both 3 and 5
 with the word "Fizz-Buzz"
}
\details{
@param n A vector of integers
 @return The modified vector of strings, containing the original integers, with appropriate integers swapped out for words.
 @examples
 fizzbuzz(1:15)

 @export
}
