% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Flagstaff_Weather.R
\docType{data}
\name{Flagstaff_Weather}
\alias{Flagstaff_Weather}
\title{A time series of daily maximum temperatures in Flagstaff, AZ from 2015 to 2019.}
\format{
a data frame with 1719 observations
\describe{
  \item{DATE}{The date of observation.}
  \item{PRCP}{Total precipitation on the observed date}
  \item{SNOW}{Snowfall}
  \item{TMAX}{Maximum temperature}
  \item{TMIN}{Minimum temperature}
}
}
\source{
\url{https://www.ncdc.noaa.gov}
}
\usage{
Flagstaff_Weather
}
\description{
A time series of daily maximum temperatures in Flagstaff, AZ from 2015 to 2019.
}
\keyword{datasets}
